+++ emulation/rocjitsu/lib/rocjitsu/src/rocjitsu/isa/arch/amdgpu/shared/execute_shared.h
@@ -3012,10 +3012,11 @@ inline void execute_v_add_i16_vop3([[maybe_unused]] Inst &inst, [[maybe_unused]]
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
-                             (static_cast<int16_t>(inst.src0.read_lane(wf, lane)) +
-                              static_cast<int16_t>(inst.src1.read_lane(wf, lane)))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
+            (static_cast<uint32_t>(static_cast<int16_t>(inst.src0.read_lane(wf, lane))) +
+             static_cast<uint32_t>(static_cast<int16_t>(inst.src1.read_lane(wf, lane))))))));
   }
 }
 
@@ -3026,9 +3027,10 @@ inline void execute_v_add_i32_vop3([[maybe_unused]] Inst &inst, [[maybe_unused]]
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         (static_cast<int32_t>(inst.src0.read_lane(wf, lane)) +
-                          static_cast<int32_t>(inst.src1.read_lane(wf, lane))));
+    inst.vdst.write_lane(
+        wf, lane,
+        (static_cast<uint32_t>(static_cast<int32_t>(inst.src0.read_lane(wf, lane))) +
+         static_cast<uint32_t>(static_cast<int32_t>(inst.src1.read_lane(wf, lane)))));
   }
 }
 
@@ -3054,10 +3056,11 @@ inline void execute_v_add_nc_i16_vop3([[maybe_unused]] Inst &inst, [[maybe_unuse
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
-                             (static_cast<int16_t>(inst.src0.read_lane(wf, lane)) +
-                              static_cast<int16_t>(inst.src1.read_lane(wf, lane)))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
+            (static_cast<uint32_t>(static_cast<int16_t>(inst.src0.read_lane(wf, lane))) +
+             static_cast<uint32_t>(static_cast<int16_t>(inst.src1.read_lane(wf, lane))))))));
   }
 }
 
@@ -3068,9 +3071,10 @@ inline void execute_v_add_nc_i32_vop3([[maybe_unused]] Inst &inst, [[maybe_unuse
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         (static_cast<int32_t>(inst.src0.read_lane(wf, lane)) +
-                          static_cast<int32_t>(inst.src1.read_lane(wf, lane))));
+    inst.vdst.write_lane(
+        wf, lane,
+        (static_cast<uint32_t>(static_cast<int32_t>(inst.src0.read_lane(wf, lane))) +
+         static_cast<uint32_t>(static_cast<int32_t>(inst.src1.read_lane(wf, lane)))));
   }
 }
 
@@ -3080,10 +3084,11 @@ inline void execute_v_add_nc_u16_vop3([[maybe_unused]] Inst &inst, [[maybe_unuse
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(
-                             (static_cast<uint16_t>(inst.src0.read_lane(wf, lane)) +
-                              static_cast<uint16_t>(inst.src1.read_lane(wf, lane))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(
+            (static_cast<uint32_t>(static_cast<uint16_t>(inst.src0.read_lane(wf, lane))) +
+             static_cast<uint32_t>(static_cast<uint16_t>(inst.src1.read_lane(wf, lane)))))));
   }
 }
 
@@ -3117,10 +3122,11 @@ inline void execute_v_add_u16_vop2([[maybe_unused]] Inst &inst, [[maybe_unused]]
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(
-                             (static_cast<uint16_t>(inst.src0.read_lane(wf, lane)) +
-                              static_cast<uint16_t>(inst.vsrc1.read_lane(wf, lane))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(
+            (static_cast<uint32_t>(static_cast<uint16_t>(inst.src0.read_lane(wf, lane))) +
+             static_cast<uint32_t>(static_cast<uint16_t>(inst.vsrc1.read_lane(wf, lane)))))));
   }
 }
 
@@ -3130,10 +3136,11 @@ inline void execute_v_add_u16_vop3([[maybe_unused]] Inst &inst, [[maybe_unused]]
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(
-                             (static_cast<uint16_t>(inst.src0.read_lane(wf, lane)) +
-                              static_cast<uint16_t>(inst.src1.read_lane(wf, lane))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(
+            (static_cast<uint32_t>(static_cast<uint16_t>(inst.src0.read_lane(wf, lane))) +
+             static_cast<uint32_t>(static_cast<uint16_t>(inst.src1.read_lane(wf, lane)))))));
   }
 }
 
@@ -13281,11 +13288,12 @@ inline void execute_v_mad_legacy_i16_vop3([[maybe_unused]] Inst &inst,
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
-                             ((static_cast<int16_t>(inst.src0.read_lane(wf, lane)) *
-                               static_cast<int16_t>(inst.src1.read_lane(wf, lane))) +
-                              static_cast<int16_t>(inst.src2.read_lane(wf, lane)))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
+            ((static_cast<uint32_t>(static_cast<int16_t>(inst.src0.read_lane(wf, lane))) *
+              static_cast<uint32_t>(static_cast<int16_t>(inst.src1.read_lane(wf, lane)))) +
+             static_cast<uint32_t>(static_cast<int16_t>(inst.src2.read_lane(wf, lane))))))));
   }
 }
 
@@ -13296,11 +13304,12 @@ inline void execute_v_mad_legacy_u16_vop3([[maybe_unused]] Inst &inst,
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(
-                             ((static_cast<uint16_t>(inst.src0.read_lane(wf, lane)) *
-                               static_cast<uint16_t>(inst.src1.read_lane(wf, lane))) +
-                              static_cast<uint16_t>(inst.src2.read_lane(wf, lane))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(
+            ((static_cast<uint32_t>(static_cast<uint16_t>(inst.src0.read_lane(wf, lane))) *
+              static_cast<uint32_t>(static_cast<uint16_t>(inst.src1.read_lane(wf, lane)))) +
+             static_cast<uint32_t>(static_cast<uint16_t>(inst.src2.read_lane(wf, lane)))))));
   }
 }
 
@@ -18474,10 +18483,11 @@ inline void execute_v_sub_i16_vop3([[maybe_unused]] Inst &inst, [[maybe_unused]]
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
-                             (static_cast<int16_t>(inst.src0.read_lane(wf, lane)) -
-                              static_cast<int16_t>(inst.src1.read_lane(wf, lane)))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
+            (static_cast<uint32_t>(static_cast<int16_t>(inst.src0.read_lane(wf, lane))) -
+             static_cast<uint32_t>(static_cast<int16_t>(inst.src1.read_lane(wf, lane))))))));
   }
 }
 
@@ -18488,9 +18498,10 @@ inline void execute_v_sub_i32_vop3([[maybe_unused]] Inst &inst, [[maybe_unused]]
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         (static_cast<int32_t>(inst.src0.read_lane(wf, lane)) -
-                          static_cast<int32_t>(inst.src1.read_lane(wf, lane))));
+    inst.vdst.write_lane(
+        wf, lane,
+        (static_cast<uint32_t>(static_cast<int32_t>(inst.src0.read_lane(wf, lane))) -
+         static_cast<uint32_t>(static_cast<int32_t>(inst.src1.read_lane(wf, lane)))));
   }
 }
 
@@ -18500,10 +18511,11 @@ inline void execute_v_sub_nc_i16_vop3([[maybe_unused]] Inst &inst, [[maybe_unuse
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
-                             (static_cast<int16_t>(inst.src0.read_lane(wf, lane)) -
-                              static_cast<int16_t>(inst.src1.read_lane(wf, lane)))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(static_cast<int16_t>(
+            (static_cast<uint32_t>(static_cast<int16_t>(inst.src0.read_lane(wf, lane))) -
+             static_cast<uint32_t>(static_cast<int16_t>(inst.src1.read_lane(wf, lane))))))));
   }
 }
 
@@ -18514,9 +18526,10 @@ inline void execute_v_sub_nc_i32_vop3([[maybe_unused]] Inst &inst, [[maybe_unuse
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         (static_cast<int32_t>(inst.src0.read_lane(wf, lane)) -
-                          static_cast<int32_t>(inst.src1.read_lane(wf, lane))));
+    inst.vdst.write_lane(
+        wf, lane,
+        (static_cast<uint32_t>(static_cast<int32_t>(inst.src0.read_lane(wf, lane))) -
+         static_cast<uint32_t>(static_cast<int32_t>(inst.src1.read_lane(wf, lane)))));
   }
 }
 
@@ -18526,10 +18539,11 @@ inline void execute_v_sub_nc_u16_vop3([[maybe_unused]] Inst &inst, [[maybe_unuse
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(
-                             (static_cast<uint16_t>(inst.src0.read_lane(wf, lane)) -
-                              static_cast<uint16_t>(inst.src1.read_lane(wf, lane))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(
+            (static_cast<uint32_t>(static_cast<uint16_t>(inst.src0.read_lane(wf, lane))) -
+             static_cast<uint32_t>(static_cast<uint16_t>(inst.src1.read_lane(wf, lane)))))));
   }
 }
 
@@ -18563,10 +18577,11 @@ inline void execute_v_sub_u16_vop2([[maybe_unused]] Inst &inst, [[maybe_unused]]
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(
-                             (static_cast<uint16_t>(inst.src0.read_lane(wf, lane)) -
-                              static_cast<uint16_t>(inst.vsrc1.read_lane(wf, lane))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(
+            (static_cast<uint32_t>(static_cast<uint16_t>(inst.src0.read_lane(wf, lane))) -
+             static_cast<uint32_t>(static_cast<uint16_t>(inst.vsrc1.read_lane(wf, lane)))))));
   }
 }
 
@@ -18576,10 +18591,11 @@ inline void execute_v_sub_u16_vop3([[maybe_unused]] Inst &inst, [[maybe_unused]]
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(
-                             (static_cast<uint16_t>(inst.src0.read_lane(wf, lane)) -
-                              static_cast<uint16_t>(inst.src1.read_lane(wf, lane))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(
+            (static_cast<uint32_t>(static_cast<uint16_t>(inst.src0.read_lane(wf, lane))) -
+             static_cast<uint32_t>(static_cast<uint16_t>(inst.src1.read_lane(wf, lane)))))));
   }
 }
 
@@ -18974,10 +18990,11 @@ inline void execute_v_subrev_u16_vop2([[maybe_unused]] Inst &inst, [[maybe_unuse
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(
-                             (static_cast<uint16_t>(inst.vsrc1.read_lane(wf, lane)) -
-                              static_cast<uint16_t>(inst.src0.read_lane(wf, lane))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(
+            (static_cast<uint32_t>(static_cast<uint16_t>(inst.vsrc1.read_lane(wf, lane))) -
+             static_cast<uint32_t>(static_cast<uint16_t>(inst.src0.read_lane(wf, lane)))))));
   }
 }
 
@@ -18988,10 +19005,11 @@ inline void execute_v_subrev_u16_vop3([[maybe_unused]] Inst &inst, [[maybe_unuse
   for (uint32_t lane = 0; lane < wf.wf_size(); ++lane) {
     if (!(exec & (1ULL << lane)))
       continue;
-    inst.vdst.write_lane(wf, lane,
-                         static_cast<uint32_t>(static_cast<uint16_t>(
-                             (static_cast<uint16_t>(inst.src1.read_lane(wf, lane)) -
-                              static_cast<uint16_t>(inst.src0.read_lane(wf, lane))))));
+    inst.vdst.write_lane(
+        wf, lane,
+        static_cast<uint32_t>(static_cast<uint16_t>(
+            (static_cast<uint32_t>(static_cast<uint16_t>(inst.src1.read_lane(wf, lane))) -
+             static_cast<uint32_t>(static_cast<uint16_t>(inst.src0.read_lane(wf, lane)))))));
   }
 }
